Move remaining Trinity logic from p2p to trinity module
