+++ p2p/constants.py
@@ -37,10 +37,6 @@ CONN_IDLE_TIMEOUT = 30
 REPLY_TIMEOUT = 3
 MAX_REQUEST_ATTEMPTS = 3
 
-# Timeout used when performing the check to ensure peers are on the same side of chain splits as
-# us.
-CHAIN_SPLIT_CHECK_TIMEOUT = 15
-
 # Default timeout before giving up on a caller-initiated interaction
 COMPLETION_TIMEOUT = 5
